Rename api-uri to messages-uri
