+++ src/lib/tsconfig.json
@@ -18,7 +18,7 @@
     "polling/poll.ts",
     "utils/hmac-sha256.ts",
     "api.ts",
-    "api-uri.ts",
+    "messages-uri.ts",
     "consts.ts",
     "contacts_service.ts",
     "interfaces/io.ts",
